+++ database/migrations/tenant/2021_10_08_205240_create_history_medical_categories_table.php
@@ -18,7 +18,7 @@ class CreateHistoryMedicalCategoriesTable extends Migration
             $table->string('name', 15);
             $table->boolean('status')->default(0);
             $table->boolean('is_various')->default(1);
-            $table->unsignedInteger('history_medical_model_id');
+            $table->unsignedBigInteger('history_medical_model_id');
             $table->timestamps();
             $table->softDeletes();
